backend/requirements.txt
fastapi==0.114.0 
pymongo==3.12.3 
python-dotenv==0.19.0
uvicorn==0.15.0
langchain-openai==0.1.25  
langchain==0.2.17 
langchain-community==0.2.19 
langchain-core==0.2.43
pydantic==2.8.1
pydantic-core==2.20.1 
requests==2.31.0
slowapi==0.1.9